refactor: update pickle vault
