+++ app/components/Vault/amplifyColumns.js
